Created ABC for Command + moved parser creation to individual command implementations
