+++ yarndevtools/commands/reviewsheetbackportupdater/review_sheet_backport_updater.py
@@ -8,8 +8,8 @@ from pythoncommons.project_utils import ProjectUtils
 
 from yarndevtools.commands.reviewsheetbackportupdater.common import ReviewSheetBackportUpdaterData
 from yarndevtools.commands.reviewsheetbackportupdater.representation import ReviewSheetBackportUpdaterOutputManager
-from yarndevtools.commands_common import CommitData
-from yarndevtools.common.shared_command_utils import SharedCommandUtils
+from yarndevtools.commands_common import CommitData, CommandAbs, GSheetArguments
+from yarndevtools.common.shared_command_utils import SharedCommandUtils, CommandType
 from yarndevtools.constants import ANY_JIRA_ID_PATTERN
 
 LOG = logging.getLogger(__name__)
@@ -55,13 +55,62 @@ class ReviewSheetBackportUpdaterConfig:
         )
 
 
-class ReviewSheetBackportUpdater:
+class ReviewSheetBackportUpdater(CommandAbs):
     def __init__(self, args, parser, output_dir: str, downstream_repo):
         self.config = ReviewSheetBackportUpdaterConfig(parser, args, output_dir)
         self.gsheet_wrapper: GSheetWrapper or None = GSheetWrapper(self.config.gsheet_options)
         self.downstream_repo = downstream_repo
         self.downstream_repo.fetch(all=True)
         self.data: ReviewSheetBackportUpdaterData = ReviewSheetBackportUpdaterData()
+
+    @staticmethod
+    def create_parser(subparsers, func_to_call):
+        parser = subparsers.add_parser(
+            CommandType.REVIEW_SHEET_BACKPORT_UPDATER.name,
+            help="Writes backport status to the Review sheet."
+            "Example: "
+            "--gsheet "
+            "--gsheet-client-secret /Users/snemeth/.secret/dummy.json "
+            "--gsheet-spreadsheet 'Failed testcases parsed from emails [generated by script]' "
+            "--gsheet-worksheet 'Failed testcases'",
+        )
+        gsheet_group = GSheetArguments.add_gsheet_arguments(parser)
+
+        gsheet_group.add_argument(
+            "--gsheet-jira-column",
+            dest="gsheet_jira_column",
+            required=False,
+            help="Name of the column that contains Jira issue IDs in the GSheet spreadsheet",
+        )
+
+        gsheet_group.add_argument(
+            "--gsheet-update-date-column",
+            dest="gsheet_update_date_column",
+            required=False,
+            help="Name of the column where this script will store last updated date in the GSheet spreadsheet",
+        )
+
+        gsheet_group.add_argument(
+            "--gsheet-status-info-column",
+            dest="gsheet_status_info_column",
+            required=False,
+            help="Name of the column where this script will store patch status info in the GSheet spreadsheet",
+        )
+
+        parser.add_argument(
+            "-v",
+            "--verbose",
+            action="store_true",
+            dest="verbose",
+            default=None,
+            required=False,
+            help="More verbose log",
+        )
+        parser.add_argument(
+            "--branches", required=True, type=str, nargs="+", help="Check backports against these branches"
+        )
+
+        parser.set_defaults(func=func_to_call)
 
     def run(self):
         LOG.info(f"Starting Review sheet backport updater. Config: \n{str(self.config)}")
