+++ yarndevtools/commands/unittestresultaggregator/unit_test_result_aggregator.py
@@ -2,7 +2,7 @@ import datetime
 import logging
 from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import List, Dict, Tuple, Set
+from typing import List, Dict, Tuple, Set, Callable
 
 from googleapiwrapper.common import ServiceType
 from googleapiwrapper.gmail_api import GmailWrapper, ThreadQueryResults
@@ -32,8 +32,10 @@ from yarndevtools.commands.unittestresultaggregator.common import (
     FailedTestCaseAggregated,
     AGGREGATED_WS_POSTFIX,
     BuildComparisonResult,
+    SummaryMode,
 )
 from yarndevtools.commands.unittestresultaggregator.representation import SummaryGenerator, UnitTestResultOutputManager
+from yarndevtools.commands_common import CommandAbs, GSheetArguments, ArgumentParserUtils
 from yarndevtools.common.shared_command_utils import CommandType
 
 VALID_OPERATION_MODES = [OperationMode.PRINT, OperationMode.GSHEET]
@@ -716,7 +718,7 @@ class TestcaseFilterResults:
         LOG.debug(f"All failed testcase objects: {self._failed_testcases}")
 
 
-class UnitTestResultAggregator:
+class UnitTestResultAggregator(CommandAbs):
     def __init__(self, args, parser, output_dir: str):
         self.config = UnitTestResultAggregatorConfig(parser, args, output_dir)
         self.testcases_to_jiras = []
@@ -735,6 +737,146 @@ class UnitTestResultAggregator:
             account_email=self.config.account_email,
         )
         self.gmail_wrapper = GmailWrapper(self.authorizer, output_basedir=self.config.email_cache_dir)
+
+    @staticmethod
+    def create_parser(subparsers, func_to_call: Callable):
+        parser = subparsers.add_parser(
+            CommandType.UNIT_TEST_RESULT_AGGREGATOR.name,
+            help="Aggregates unit test results from a gmail account."
+            "Example: "
+            "--gsheet "
+            "--gsheet-client-secret /Users/snemeth/.secret/dummy.json "
+            "--gsheet-spreadsheet 'Failed testcases parsed from emails [generated by script]' "
+            "--gsheet-worksheet 'Failed testcases'",
+        )
+        gsheet_group = GSheetArguments.add_gsheet_arguments(parser)
+
+        gsheet_group.add_argument(
+            "--gsheet-compare-with-jira-table",
+            dest="gsheet_compare_with_jira_table",
+            type=str,
+            help="This should be provided if comparison of failed testcases with reported jira table must be performed. "
+            "The value is a name to a worksheet, for example 'testcases with jiras'.",
+        )
+
+        parser.add_argument(
+            "--account-email",
+            required=True,
+            type=str,
+            help="Email address of Gmail account that will be used to Gmail API authentication and fetching data.",
+        )
+
+        parser.add_argument(
+            "-q",
+            "--gmail-query",
+            required=True,
+            type=str,
+            help="Gmail query string that will be used to get emails to parse.",
+        )
+
+        parser.add_argument(
+            "--smart-subject-query",
+            action="store_true",
+            default=False,
+            help="Whether to fix Gmail queries like: 'Subject: YARN Daily unit test report', "
+            "where the subject should have been between quotes.",
+        )
+
+        parser.add_argument(
+            "-v",
+            "--verbose",
+            action="store_true",
+            dest="verbose",
+            default=None,
+            required=False,
+            help="More verbose log",
+        )
+
+        parser.add_argument(
+            "-m",
+            "--match-expression",
+            required=False,
+            # TODO
+            type=ArgumentParserUtils.matches_match_expression_pattern,
+            nargs="+",
+            help="Line matcher expression, this will be converted to a regex. "
+            "For example, if expression is org.apache, the regex will be .*org\\.apache\\.* "
+            "Only lines in the mail content matching for this expression will be considered as a valid line.",
+        )
+
+        parser.add_argument(
+            "-s",
+            "--skip-lines-starting-with",
+            required=False,
+            type=str,
+            nargs="+",
+            help="If lines starting with these strings, they will not be considered as a line to parse",
+        )
+
+        parser.add_argument(
+            "-l",
+            "--request-limit",
+            dest="request_limit",
+            type=int,
+            help="Limit the number of API requests",
+        )
+
+        parser.add_argument("--email-content-line-separator", type=str, default=DEFAULT_LINE_SEP)
+
+        parser.add_argument(
+            "--truncate-subject",
+            dest="truncate_subject",
+            type=str,
+            help="Whether to truncate subject in outputs. The specified string will be cropped "
+            "from the full value of subject strings when printing them to any destination.",
+        )
+
+        parser.add_argument(
+            "--abbreviate-testcase-package",
+            dest="abbrev_testcase_package",
+            type=str,
+            help="Whether to abbreviate testcase package names in outputs in order to screen space. "
+            "The specified string will be abbreviated with the starting letters."
+            "For example, specifying 'org.apache.hadoop.yarn' will be converted to 'o.a.h.y' "
+            "when printing testcase names to any destination.",
+        )
+
+        parser.add_argument(
+            "--summary-mode",
+            dest="summary_mode",
+            type=str,
+            choices=[sm.value for sm in SummaryMode],
+            default=SummaryMode.HTML.value,
+            help="Summary file(s) will be written in this mode. Defaults to HTML.",
+        )
+
+        parser.add_argument(
+            "--aggregate-filters",
+            dest="aggregate_filters",
+            required=True,
+            type=str,
+            nargs="+",
+            help="Execute some post filters on the email results. "
+            "The resulted emails and testcases for each filter will be aggregated to "
+            "a separate worksheet with name <WS>_aggregated_<aggregate-filter> where WS is equal to the "
+            "value specified by the --gsheet-worksheet argument.",
+        )
+
+        exclusive_group = parser.add_mutually_exclusive_group(required=True)
+        exclusive_group.add_argument(
+            "-p", "--print", action="store_true", dest="do_print", help="Print results to console", required=False
+        )
+        exclusive_group.add_argument(
+            "-g",
+            "--gsheet",
+            action="store_true",
+            dest="gsheet",
+            default=False,
+            required=False,
+            help="Export values to Google sheet. Additional gsheet arguments need to be specified!",
+        )
+
+        parser.set_defaults(func=func_to_call)
 
     def _load_and_convert_known_test_failures_in_jira(self):
         raw_data_from_gsheet = self.gsheet_wrapper.read_data(self.config.gsheet_jira_table, "A1:E150")
