+++ yarndevtools/commands/unittestresultfetcher/unit_test_result_fetcher.py
@@ -7,7 +7,7 @@ import time
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from enum import Enum
-from typing import List, Dict, Set, Tuple, Any
+from typing import List, Dict, Set, Tuple, Any, Callable
 
 from googleapiwrapper.common import ServiceType
 from googleapiwrapper.google_auth import GoogleApiAuthorizer
@@ -33,6 +33,7 @@ from pythoncommons.project_utils import PROJECTS_BASEDIR_NAME
 from pythoncommons.string_utils import auto_str
 
 from yarndevtools.cdsw.constants import SECRET_PROJECTS_DIR
+from yarndevtools.commands_common import CommandAbs, EmailArguments
 from yarndevtools.common.shared_command_utils import FullEmailConfig, CommandType
 
 from yarndevtools.constants import YARNDEVTOOLS_MODULE_NAME
@@ -833,13 +834,174 @@ class UnitTestResultFetcherConfig:
 
 
 # TODO Separate all download functionality: Progress of downloads, code that fetches API, etc.
-class UnitTestResultFetcher:
+class UnitTestResultFetcher(CommandAbs):
     def __init__(self, args, output_dir):
         self.config = UnitTestResultFetcherConfig(output_dir, args)
         self.reports: Dict[str, JenkinsJobReport] = {}  # key is the Jenkins job name
         self.cache: Cache = self._create_cache(self.config)
         self.email: Email = Email(self.config.email)
         self.sent_requests: int = 0
+
+    @staticmethod
+    def create_parser(subparsers, func_to_call: Callable):
+        parser = subparsers.add_parser(
+            CommandType.UNIT_TEST_RESULT_FETCHER.name,
+            help="Fetches, parses and sends unit test result reports from Jenkins in email."
+            "Example: "
+            "--mode jenkins_master "
+            "--jenkins-url {jenkins_base_url} "
+            "--job-names {job_names} "
+            "--testcase-filter org.apache.hadoop.yarn "
+            "--smtp_server smtp.gmail.com "
+            "--smtp_port 465 "
+            "--account_user someuser@somemail.com "
+            "--account_password somepassword "
+            "--sender 'YARN jenkins test reporter' "
+            "--recipients snemeth@cloudera.com "
+            "--testcase-filter YARN:org.apache.hadoop.yarn MAPREDUCE:org.apache.hadoop.mapreduce HDFS:org.apache.hadoop.hdfs "
+            "--num-builds jenkins_examine_unlimited_builds "
+            "--omit-job-summary "
+            "--download-uncached-job-data",
+        )
+        EmailArguments.add_email_arguments(parser, add_subject=False, add_attachment_filename=False)
+
+        parser.add_argument(
+            "--omit-job-summary",
+            action="store_true",
+            default=False,
+            help="Do not print job summaries to the console or the log file",
+        )
+
+        parser.add_argument(
+            "--force-download-jobs",
+            action="store_true",
+            dest="force_download_mode",
+            help="Force downloading data from all builds. "
+            "If this is set to true, all job data will be downloaded, regardless if they are already in the cache",
+        )
+
+        parser.add_argument(
+            "--download-uncached-job-data",
+            action="store_true",
+            dest="download_uncached_job_data",
+            help="Download data for all builds that are not in cache yet or was removed from the cache, for any reason.",
+        )
+
+        parser.add_argument(
+            "--force-sending-email",
+            action="store_true",
+            dest="force_send_email",
+            help="Force sending email report for all builds.",
+        )
+
+        parser.add_argument(
+            "-s",
+            "--skip-sending-email",
+            dest="skip_email",
+            type=bool,
+            help="Skip sending email report for all builds.",
+        )
+
+        parser.add_argument(
+            "--reset-sent-state-for-jobs",
+            nargs="+",
+            type=str,
+            dest="reset_sent_state_for_jobs",
+            default=[],
+            help="Reset email sent state for these jobs.",
+        )
+
+        parser.add_argument(
+            "--reset-job-build-data-for-jobs",
+            nargs="+",
+            type=str,
+            dest="reset_job_build_data_for_jobs",
+            default=[],
+            help="Reset job build data for these jobs. Useful when job build data is corrupted.",
+        )
+
+        parser.add_argument(
+            "-m",
+            "--mode",
+            type=str,
+            dest="jenkins_mode",
+            choices=[m.mode_name.lower() for m in UnitTestResultFetcherMode],
+            help="Jenkins mode. Used to pre-configure --jenkins-url and --job-names. "
+            "Will take precendence over URL and job names, if they are also specified!",
+        )
+
+        parser.add_argument(
+            "-J",
+            "--jenkins-url",
+            type=str,
+            dest="jenkins_url",
+            help="Jenkins URL to fetch results from",
+            default="http://build.infra.cloudera.com/",
+        )
+        parser.add_argument(
+            "-j",
+            "--job-names",
+            type=str,
+            dest="job_names",
+            help="Jenkins job name to fetch results from",
+            default="Mawo-UT-hadoop-CDPD-7.x",
+        )
+
+        # TODO Rationalize this vs. request-limit:
+        #  Num builds is intended to be used for determining to process the builds that are not yet processed / sent in mail
+        #  Request limit is to limit the number of builds processed for each Jenkins job
+        parser.add_argument(
+            "-n",
+            "--num-builds",
+            type=str,
+            dest="num_builds",
+            help="Number of days of Jenkins jobs to examine. "
+            "Special value of 'jenkins_examine_unlimited_builds' will examine all unknown builds.",
+            default="14",
+        )
+        parser.add_argument(
+            "-rl",
+            "--request-limit",
+            type=int,
+            dest="req_limit",
+            help="Request limit",
+            default=999,
+        )
+
+        def tc_filter_validator(value):
+            strval = str(value)
+            if ":" not in strval:
+                raise ValueError("Filter specification should be in this format: '<project>:<filter statement>'")
+            return strval
+
+        parser.add_argument(
+            "-t",
+            "--testcase-filter",
+            dest="tc_filters",
+            nargs="+",
+            type=tc_filter_validator,
+            help="Testcase filters in format: <project:filter statement>",
+        )
+
+        # TODO change this to disable cache
+        parser.add_argument(
+            "-d",
+            "--disable-file-cache",
+            dest="disable_file_cache",
+            type=bool,
+            help="Whether to disable Jenkins report file cache",
+        )
+
+        parser.add_argument(
+            "-ct",
+            "--cache-type",
+            type=str,
+            dest="cache_type",
+            choices=[ct.name.lower() for ct in UnitTestResultFetcherCacheType],
+            help="The type of the cache. Either file or google_drive",
+        )
+
+        parser.set_defaults(func=func_to_call)
 
     @staticmethod
     def _convert_to_cache_build_key(failed_build: FailedJenkinsBuild):
